Initial attempt to port nix to 9front.
Rename Ureg.ip to Ureg.pc
Move some parameters from ulong to usize
mkfile for boot
nix script for initial binds.

+++ sys/src/boot/pc/lib.h
@@ -8,9 +8,9 @@
  * mem routines
  */
 extern	void*	memccpy(void*, void*, int, ulong);
-extern	void*	memset(void*, int, ulong);
-extern	int	memcmp(void*, void*, ulong);
-extern	void*	memmove(void*, void*, ulong);
+extern	void*	memset(void*, int, usize);
+extern	int	memcmp(void*, void*, usize);
+extern	void*	memmove(void*, void*, usize);
 extern	void*	memchr(void*, int, ulong);
 
 /*
